Refactored draw. Update still doesn't work because of other variables in drummachine.js. Working on it.
