+++ public/js/drummachine.js
@@ -1,5 +1,6 @@
 const beatMod = require('./beat')
 const synthMod = require('./synthesis')
+const drawMod = require('./draw.js')
 
 // Events
 // init() once the page has finished loading.
@@ -21,7 +22,6 @@ var effectWetMix = 1.0;
 var timeoutId;
 
 var startTime;
-var lastDrawTime = -1;
 
 var kits;
 
@@ -44,6 +44,8 @@ var rhythmIndex = 0;
 var noteTime = 0.0;
 
 var instruments = ['Kick', 'Snare', 'HiHat', 'Tom1', 'Tom2', 'Tom3'];
+exports.instruments = instruments //TODO, REFACTOR THIS TO SOMEWHERE
+
 
 var volumes = [0, 0.3, 1];
 
@@ -594,9 +596,9 @@ function schedule() {
 
 
         // Attempt to synchronize drawing time with sound
-        if (noteTime != lastDrawTime) {
-            lastDrawTime = noteTime;
-            drawPlayhead((rhythmIndex + 15) % 16);
+        if (noteTime != drawMod.lastDrawTime) {
+            drawMod.setLastDrawTime(noteTime);
+            drawMod.drawPlayhead((rhythmIndex + 15) % 16);
         }
 
         advanceNote();
@@ -789,7 +791,7 @@ function handleButtonMouseDown(event) {
     if (instrumentIndex == currentlyActiveInstrument)
         showCorrectNote( rhythmIndex, notes[rhythmIndex] );
 
-    drawNote(notes[rhythmIndex], rhythmIndex, instrumentIndex);
+    drawMod.drawNote(notes[rhythmIndex], rhythmIndex, instrumentIndex);
 
     if (newNoteValue) {
         switch(instrumentIndex) {
@@ -1083,7 +1085,7 @@ function updateControls() {
                 case 5: notes = beatMod.theBeat.rhythm6; break;
             }
 
-            drawNote(notes[i], i, j);
+            drawMod.drawNote(notes[i], i, j);
         }
     }
 
@@ -1100,40 +1102,6 @@ function updateControls() {
     sliderSetPosition('tom3_thumb', beatMod.theBeat.tom3PitchVal);
 }
 
-function drawNote(draw, xindex, yindex) {
-    var elButton = document.getElementById(instruments[yindex] + '_' + xindex);
-    switch (draw) {
-        case 0: elButton.src = 'images/button_off.png'; break;
-        case 1: elButton.src = 'images/button_half.png'; break;
-        case 2: elButton.src = 'images/button_on.png'; break;
-    }
-}
-
-function redrawAllNotes() {
-    for (y = 0; y < 6; y++) { //6 rhythm patterns in theBeat
-        for (x = 0; x < 16; x++)  { //16 beat subdivisions
-            if(x >= beatMod.theBeat['rhythm'+(y+1).toString()].length){
-                drawNote(0, x, y);
-            }
-            else {
-                drawNote(beatMod.theBeat['rhythm'+(y+1).toString()][x], x, y);
-            }
-        }
-    }
-}
-
-function drawPlayhead(xindex) {
-    var lastIndex = (xindex + 15) % 16;
-
-    var elNew = document.getElementById('LED_' + xindex);
-    var elOld = document.getElementById('LED_' + lastIndex);
-
-    elNew.src = 'images/LED_on.png';
-    elOld.src = 'images/LED_off.png';
-
-    hideBeat( lastIndex );
-    showBeat( xindex );
-}
 
 function filterFrequencyFromCutoff( cutoff ) {
     var nyquist = 0.5 * context.sampleRate;
